refactor: 접두사 변경 및 함수 분리, validator(유효성 검증) view 폴더로 이동
